Extract raw query metadata
